+++ src/stream/src/executor/hash_agg.rs
@@ -15,7 +15,6 @@
 use std::collections::{HashMap, HashSet};
 use std::marker::PhantomData;
 use std::ptr::NonNull;
-use std::sync::atomic::AtomicU64;
 use std::sync::Arc;
 
 use futures::{stream, StreamExt, TryStreamExt};
@@ -24,8 +23,8 @@ use iter_chunks::IterChunks;
 use itertools::Itertools;
 use risingwave_common::array::StreamChunk;
 use risingwave_common::buffer::{Bitmap, BitmapBuilder};
-use risingwave_common::catalog::{Schema, TableId};
-use risingwave_common::hash::{HashKey, PrecomputedBuildHasher, SerializedKey};
+use risingwave_common::catalog::Schema;
+use risingwave_common::hash::{HashKey, PrecomputedBuildHasher};
 use risingwave_common::util::epoch::EpochPair;
 use risingwave_common::util::iter_util::ZipEqFast;
 use risingwave_storage::StateStore;
@@ -45,8 +44,7 @@ use crate::error::StreamResult;
 use crate::executor::aggregation::{generate_agg_schema, AggCall, AggGroup as GenericAggGroup};
 use crate::executor::error::StreamExecutorError;
 use crate::executor::monitor::StreamingMetrics;
-use crate::executor::test_utils::agg_executor::{create_agg_state_storage, create_result_table};
-use crate::executor::{ActorContext, BoxedMessageStream, Executor, Message, PkIndices};
+use crate::executor::{BoxedMessageStream, Executor, Message};
 use crate::task::AtomicU64Ref;
 
 type AggGroup<S> = GenericAggGroup<S, OnlyOutputIfHasInput>;
@@ -594,67 +592,6 @@ impl<K: HashKey, S: StateStore> HashAggExecutor<K, S> {
     }
 }
 
-/// NOTE(kwannoel): This should only be used by `test` or `bench`.
-#[allow(clippy::too_many_arguments)]
-pub async fn new_boxed_hash_agg_executor<S: StateStore>(
-    store: S,
-    input: Box<dyn Executor>,
-    agg_calls: Vec<AggCall>,
-    row_count_index: usize,
-    group_key_indices: Vec<usize>,
-    pk_indices: PkIndices,
-    extreme_cache_size: usize,
-    executor_id: u64,
-) -> Box<dyn Executor> {
-    let mut storages = Vec::with_capacity(agg_calls.iter().len());
-    for (idx, agg_call) in agg_calls.iter().enumerate() {
-        storages.push(
-            create_agg_state_storage(
-                store.clone(),
-                TableId::new(idx as u32),
-                agg_call,
-                &group_key_indices,
-                &pk_indices,
-                input.as_ref(),
-            )
-            .await,
-        )
-    }
-
-    let result_table = create_result_table(
-        store,
-        TableId::new(agg_calls.len() as u32),
-        &agg_calls,
-        &group_key_indices,
-        input.as_ref(),
-    )
-    .await;
-
-    HashAggExecutor::<SerializedKey, S>::new(AggExecutorArgs {
-        input,
-        actor_ctx: ActorContext::create(123),
-        pk_indices,
-        executor_id,
-
-        extreme_cache_size,
-
-        agg_calls,
-        row_count_index,
-        storages,
-        result_table,
-        distinct_dedup_tables: Default::default(),
-        watermark_epoch: Arc::new(AtomicU64::new(0)),
-
-        extra: GroupAggExecutorExtraArgs {
-            group_key_indices,
-            chunk_size: 1024,
-            metrics: Arc::new(StreamingMetrics::unused()),
-        },
-    })
-    .unwrap()
-    .boxed()
-}
-
 #[cfg(test)]
 pub mod tests {
 
@@ -672,7 +609,7 @@ pub mod tests {
     use risingwave_storage::StateStore;
 
     use crate::executor::aggregation::{AggArgs, AggCall};
-    use crate::executor::hash_agg::new_boxed_hash_agg_executor;
+    use crate::executor::test_utils::agg_executor::new_boxed_hash_agg_executor;
     use crate::executor::test_utils::*;
     use crate::executor::{Message, PkIndices};
