+++ src/stream/src/executor/test_utils.rs
@@ -227,6 +227,7 @@ pub mod agg_executor {
     use std::sync::Arc;
 
     use risingwave_common::catalog::{ColumnDesc, ColumnId, TableId};
+    use risingwave_common::hash::SerializedKey;
     use risingwave_common::types::DataType;
     use risingwave_common::util::sort_util::OrderType;
     use risingwave_expr::expr::AggKind;
@@ -234,10 +235,14 @@ pub mod agg_executor {
 
     use crate::common::table::state_table::StateTable;
     use crate::common::StateTableColumnMapping;
-    use crate::executor::agg_common::{AggExecutorArgs, SimpleAggExecutorExtraArgs};
+    use crate::executor::agg_common::{
+        AggExecutorArgs, GroupAggExecutorExtraArgs, SimpleAggExecutorExtraArgs,
+    };
     use crate::executor::aggregation::{AggCall, AggStateStorage};
+    use crate::executor::monitor::StreamingMetrics;
     use crate::executor::{
-        ActorContextRef, BoxedExecutor, Executor, GlobalSimpleAggExecutor, PkIndices,
+        ActorContext, ActorContextRef, BoxedExecutor, Executor, GlobalSimpleAggExecutor,
+        HashAggExecutor, PkIndices,
     };
 
     /// Create state storage for the given agg call.
@@ -347,6 +352,67 @@ pub mod agg_executor {
             (0..group_key_indices.len()).collect(),
         )
         .await
+    }
+
+    /// NOTE(kwannoel): This should only be used by `test` or `bench`.
+    #[allow(clippy::too_many_arguments)]
+    pub async fn new_boxed_hash_agg_executor<S: StateStore>(
+        store: S,
+        input: Box<dyn Executor>,
+        agg_calls: Vec<AggCall>,
+        row_count_index: usize,
+        group_key_indices: Vec<usize>,
+        pk_indices: PkIndices,
+        extreme_cache_size: usize,
+        executor_id: u64,
+    ) -> Box<dyn Executor> {
+        let mut storages = Vec::with_capacity(agg_calls.iter().len());
+        for (idx, agg_call) in agg_calls.iter().enumerate() {
+            storages.push(
+                create_agg_state_storage(
+                    store.clone(),
+                    TableId::new(idx as u32),
+                    agg_call,
+                    &group_key_indices,
+                    &pk_indices,
+                    input.as_ref(),
+                )
+                .await,
+            )
+        }
+
+        let result_table = create_result_table(
+            store,
+            TableId::new(agg_calls.len() as u32),
+            &agg_calls,
+            &group_key_indices,
+            input.as_ref(),
+        )
+        .await;
+
+        HashAggExecutor::<SerializedKey, S>::new(AggExecutorArgs {
+            input,
+            actor_ctx: ActorContext::create(123),
+            pk_indices,
+            executor_id,
+
+            extreme_cache_size,
+
+            agg_calls,
+            row_count_index,
+            storages,
+            result_table,
+            distinct_dedup_tables: Default::default(),
+            watermark_epoch: Arc::new(AtomicU64::new(0)),
+
+            extra: GroupAggExecutorExtraArgs {
+                group_key_indices,
+                chunk_size: 1024,
+                metrics: Arc::new(StreamingMetrics::unused()),
+            },
+        })
+        .unwrap()
+        .boxed()
     }
 
     pub async fn new_boxed_simple_agg_executor<S: StateStore>(
